adding grpc

+++ domain/src/main/java/com/example/domain/model/CancerVolume.kt
@@ -1,7 +1,7 @@
 package com.example.domain.model
 
 class CancerVolume (
-    var volume : Int,
+    var volume : Float,
     var sequence: MRISequence,
     var affinityMatrix: Array<Array<FloatArray>>,
 )
